fix: Resolve more Windows build issues
  - Guard cxxabi.h and sys/file.h includes for non-Windows platforms only
  - Use strerror_s instead of strerror on Windows to avoid deprecation warning
  - Add explicit cast for size_t to unsigned long conversion to avoid data loss warning
  - These headers and functions are POSIX/GCC specific and not available on MSVC

+++ core/src/main/cpp/src/typemgr.h
@@ -35,7 +35,7 @@ namespace xtree {
 
     template<>
     struct xtTypeTraits<string> {
-        static long unsigned int length(string value) { return value.length(); }
+        static long unsigned int length(string value) { return static_cast<long unsigned int>(value.length()); }
         static void setValue(string rhs, string &lhs) { lhs = rhs;  }
         /*static int compare(string v1, string v2) {
